+++ CHANGELOG.md
@@ -9,9 +9,15 @@ this changelog also includes the original WARDEN changelog.
 This release introduces breaking changes to the integrated ("Supreme") components to deliver **truly, fully integrated
 key and app attestation**, pinning down the last unnecessarily moving parts.
 
-* Ignore Android leaf cert validity by default, as the default Warden Supreme Verifier uses strong cryptographic nonces.
-* Rename Warden -> Makoto to more clearly distinguish individual components by name
+* Rename `Warden` -> `Makoto` to more clearly distinguish individual components by name
     * A `typealias Warden = Makoto` is present, but marked as deprecated for
+* Rename `AndroidAttestationChecker` -> `Roboto` to more clearly distinguish individual components by name
+    * Rename `HardwareAttestationChecker` -> `HardwareAttestationVerifier` (and introduce typealias, but marked as deprecated) 
+    * Rename `NougatHybridAttestationChecker` -> `NougatHybridAttestationVerifier` (and introduce typealias, but marked as deprecated) 
+    * Rename `SoftwareAttestationChecker` -> `SoftwareAttestationVerifier` (and introduce typealias, but marked as deprecated) 
+* Ignore Android leaf cert validity by default, because Warden Supreme, by default, uses random cryptographic nonces.
+* Android `attestationStatementValiditySeconds` defaults to `null`, because Warden Supreme, by default, uses random cryptographic nonces.
+* Per-App StrongboxOverride
 * Rework Trust Anchor Management:
     * Introduce `TrustedRoot` interface to represent trust anchors
         * `TrustedRoot.Certificate` for certificates